+++ src/organisms/CreateProject/Layout/ClientDetails/ClientDetails.tsx
@@ -1,10 +1,8 @@
 import { Typography, TextField } from "@mui/material";
-import { useFormik } from "formik";
-
 import { ButtonSection, Container, FlexRow } from "./ClientDetails.style";
 import Button from "../../../../atoms/Button/Button";
-import { clientDetailsSchema } from "../../Schema";
 import SelectOption from "../../../../atoms/Select/SelectOption";
+import useClientDetailsForm from "./ClientDetails.hook"; // Import the custom hook
 
 const ClientDetails = ({
   data,
@@ -38,39 +36,12 @@ const ClientDetails = ({
   handleNext: () => void;
   handlePrevious: () => void;
 }) => {
-  const formik = useFormik({
-    initialValues: {
-      clientName: data.clientName || "",
-      clientEmail: data.clientEmail || "",
-      phoneNumber: data.phoneNumber || "",
-      city: data.city || "",
-      country: data.country || "",
-      state: data.state || "",
-      zipCode: data.zipCode || "",
-      addressLine1: data.addressLine1 || "",
-      addressLine2: data.addressLine2 || "",
-    },
-    validationSchema: clientDetailsSchema,
-    validateOnBlur: true,
-    onSubmit: (values) => {
-      updateData(values);
-      handleNext();
-      console.log(values);
-    },
+  const { formik, isFormValid } = useClientDetailsForm({
+    data,
+    updateData,
+    handleNext,
+    handlePrevious,
   });
-
-  const isFormValid = () => {
-    return (
-      formik.values.clientName &&
-      formik.values.clientEmail &&
-      formik.values.phoneNumber &&
-      formik.values.city &&
-      formik.values.country &&
-      formik.values.state &&
-      formik.values.zipCode &&
-      formik.values.addressLine1
-    );
-  };
 
   return (
     <Container>
@@ -174,7 +145,6 @@ const ClientDetails = ({
           variant="outlined"
           name="zipCode"
           type="number"
-          placeholder="ex. 098098"
           value={formik.values.zipCode}
           onChange={formik.handleChange}
           onBlur={formik.handleBlur}
@@ -218,7 +188,7 @@ const ClientDetails = ({
           color="primary"
           variant="contained"
           disabled={!isFormValid()}
-          onClick={formik.handleSubmit}
+          onClick={() => formik.handleSubmit()}
         />
       </ButtonSection>
     </Container>
